requirements.txt
streamlit
google-cloud-aiplatform>=1.38.1
